integration_app/requirements.txt
requests==2.25.1
Flask==1.0.2
gunicorn
google-cloud-storage==1.42.3
